ロジック部分をhooksに分離
