+++ src/app/user/[account_id]/page.tsx
@@ -1,74 +1,13 @@
 "use client";
 
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Image from "next/image";
 import styles from "./page.module.css";
-import { supabase } from "@/supabase/supabase";
-import type User from "@/types/User";
-import Technology from "@/types/Technology";
+import useShowUser from "./useShowUser";
 
 const UserPage = ({ params }: { params: { account_id: string } }) => {
   const accountID = params.account_id;
-  const [userData, setUserData] = useState<User | null>(null);
-  const [userIconUrl, setUserIconUrl] = useState<string | null>(null);
-  const [loading, setLoading] = useState(true);
-  console.log(userData);
-
-  useEffect(() => {
-    const fetchUserIcons = async (id: string) => {
-      const extensions = ["JPG", "jpg", "jpeg", "png", "gif"];
-      let imageUrl = null;
-
-      for (const ext of extensions) {
-        const { data } = supabase.storage
-          .from("user_icons")
-          .getPublicUrl(`${id}.${ext}`);
-        if (data?.publicUrl) {
-          imageUrl = data.publicUrl;
-          break;
-        }
-      }
-      return imageUrl;
-    };
-
-    const fetchUserData = async () => {
-      const { data, error } = await supabase
-        .from("users")
-        .select(
-          `
-          *,
-          users_technologies(technology_id(id, name))
-          `
-        )
-        .eq("account_id", accountID)
-        .single();
-
-      if (error) {
-        console.error("Error fetching user data:", error);
-        setUserData(null);
-      } else {
-        setUserData({
-          id: data?.id,
-          accountID: data?.account_id,
-          name: data?.name,
-          nickname: data?.nickname,
-          introduction: data?.introduction || "",
-          technologies:
-            data?.users_technologies.map(
-              (users_technology: { technology_id: Technology }) =>
-                users_technology.technology_id
-            ) || [],
-        } as User);
-
-        const fetchedURL = await fetchUserIcons(data?.id);
-        setUserIconUrl(fetchedURL);
-      }
-
-      setLoading(false);
-    };
-
-    fetchUserData();
-  }, []);
+  const { userData, userIconUrl, loading } = useShowUser(accountID);
 
   if (loading) {
     return (
